feat: health check for updaters (#207)
This is a pretty hacky solution to restart the app after one of the
updaters crashed.

It's far from ideal, but I'm not sure why the contract updater panicked
before. As Fly.io stores logs for only a couple of hours, it's hard to
detect the issue. This will allow us to get back to work and luckily to
identify the issue once we get notification and get some logging.